Rename profile-2 to profile
