+++ src/applications/personalization/profile/sass/profile.scss
