+++ src/applications/personalization/profile/tests/unit-test-helpers.js
@@ -1,7 +1,7 @@
 import { renderInReduxProvider } from 'platform/testing/unit/react-testing-library-helpers';
 
 import profile from 'applications/personalization/profile360/reducers';
-import connectedApps from 'applications/personalization/profile-2/components/connected-apps/reducers/connectedApps';
+import connectedApps from 'applications/personalization/profile/components/connected-apps/reducers/connectedApps';
 
 export function wait(timeout) {
   return new Promise(resolve => {
